refactor(P4a): mirror/catchUp 引擎收敛为 src/server/mirror-engine.js
15 个 mirror*/catchUp* 模块级可变量与整套编排（catchUpTickOnce/setMirror/catchUpTick/
rescheduleCatchUp 等）从 app.js 顶层搬入自持模块，对 app.js 只暴露窄接口：
start/stop/catchUpTick/mirrorOwnedBy/isReadonly/clearMirrorOnViewChange/
requestRebaseline/takeOver/snapshot。

★纯位置迁移，逻辑一行未改：规范化 diff 验证 361 行原文与搬迁后仅 1 行差异
（statusBridgeOff 改为注入参数）。

实测订正了计划书两处注入面结论：registryBusy 是局部变量非模块级符号；
viewingCwd/broadcastInstances/instancesPayload 在引擎内根本未被引用。
真实注入面 9 项：io/agents/instanceState/getViewingInstanceId/viewingCwdOf/
serviceStartedAt/scheduleStatusRefresh/readCliSnapshotForSession/statusBridgeOff。

保留的既有契约：重连快照 payload 形状与 setMirror 广播有意不同（不含
quietTicks/remainingMs/cliSeen）；接管判定读全局 mirrorReadonly 而非归属判定。

app.js 3259→2920 行、顶层 let 减 15 个。check 全绿、npm test 零新增失败
（含集成测试「连接时总是收到权威 mirror_state」）、E2E 218 全过。
